Add amiri-ttf-1.000
Amiri (أميري) is a classical Arabic typeface in Naskh style for
typesetting books and other running text.

Amiri is a revival of the beautiful typeface pioneered in early 20th
century by Bulaq Press in Cairo, also known as Amiria Press, after which
the font is named.

The uniqueness of this typeface comes from its superb balance between the
beauty of Naskh calligraphy on one hand, the constraints and requirements
of elegant typography on the other. Also, it is one of the few metal
typefaces that were used in typesetting the Koran, making it a good
source for a digital typeface to be used in typesetting Koranic verses.